refactor: modularize summarizer, prompts config
